renamed moved everything belonging to the main screen and its view models to ui.main package
