+++ app/src/main/java/com/peternaggschga/gwent/MainActivity.java
@@ -14,7 +14,9 @@ import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.lifecycle.ViewModelProvider;
 
 import com.peternaggschga.gwent.data.RowType;
-import com.peternaggschga.gwent.ui.GameBoardViewModel;
+import com.peternaggschga.gwent.ui.main.GameBoardViewModel;
+import com.peternaggschga.gwent.ui.main.MenuUiStateObserver;
+import com.peternaggschga.gwent.ui.main.RowUiStateObserver;
 
 import io.reactivex.rxjava3.schedulers.Schedulers;
