+++ app/src/main/java/com/peternaggschga/gwent/domain/cases/RowStateUseCase.java
@@ -6,7 +6,7 @@ import com.peternaggschga.gwent.data.RowType;
 import com.peternaggschga.gwent.data.UnitEntity;
 import com.peternaggschga.gwent.data.UnitRepository;
 import com.peternaggschga.gwent.domain.damage.DamageCalculator;
-import com.peternaggschga.gwent.ui.RowUiState;
+import com.peternaggschga.gwent.ui.main.RowUiState;
 
 import java.util.Collection;
