+++ app/src/main/java/com/peternaggschga/gwent/ui/main/RowUiState.java
@@ -1,4 +1,4 @@
-package com.peternaggschga.gwent.ui;
+package com.peternaggschga.gwent.ui.main;
 
 import static org.valid4j.Assertive.require;
